* fix two warnings in exceptionhandler.c  * add Q_ and Qn_ gettext macros (copied from Freeciv: GPL)  * remove hack for ambiguous translations from droid.c; use Q_ instead  * update makefiles accordingly  * `make -C po update-po`
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@1417 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/framework/exceptionhandler.c
@@ -73,14 +73,14 @@ static LONG WINAPI windowsExceptionHandler(PEXCEPTION_POINTERS pExceptionInfo)
 			}
 			else
 			{
-				sprintf( resultMessage, "Failed to save dump file to '%s' (error %d)", miniDumpPath, GetLastError() );
+				sprintf( resultMessage, "Failed to save dump file to '%s' (error %d)", miniDumpPath, (int)GetLastError() );
 				MessageBoxA( NULL, resultMessage, applicationName, MB_OK );
 			}
 			CloseHandle(miniDumpFile);
 		}
 		else
 		{
-			sprintf( resultMessage, "Failed to create dump file '%s' (error %d)", miniDumpPath, GetLastError() );
+			sprintf( resultMessage, "Failed to create dump file '%s' (error %d)", miniDumpPath, (int)GetLastError() );
 			MessageBoxA( NULL, resultMessage, applicationName, MB_OK );
 		}
 	}
